split the large rate.py into two files
