+++ pynucastro/rates/__init__.py
@@ -1,5 +1,6 @@
 """The core routines needed to read and interpret nuclear reaction rates"""
 
-__all__ = ["rate"]
+__all__ = ["rate", "library"]
 
-from .rate import Tfactors, Library, ReacLibLibrary, Rate, RateFilter, UnsupportedNucleus, Nucleus, list_known_rates
+from .rate import Tfactors, Rate, RateFilter, UnsupportedNucleus, Nucleus, list_known_rates
+from .library import Library, ReacLibLibrary
